wayland: introduce poxy wrappers for multi threads to queue events
Queue events with multi threads are prone to race conditions. The
problem is that one thread can read and queue events after another
thread creates a proxy but before it sets the queue.

This may result in the event to the proxy being silently dropped, or
potentially dispatched on the wrong thread had the creating thread set
the implementation before setting the queue.

The patch introduce proxy wrappers to fix the problem. A proxy wrapper
is a wl_proxy struct that will never itself proxy any events, but may be
used by the client to set a queue, and use it instead of the original proxy
when sending requests that creates new proxies.

The patch is from wayland 1.11. The details can refer
https://cgit.freedesktop.org/wayland/wayland/commit/?id=6d29c0da3cd168e08187cd204d2314188479c0f1
https://cgit.freedesktop.org/wayland/wayland/commit/?id=6fe12f02e3b4879cd3d5faa08f023cc761d13be9

Change-Id: I9d35dd472a1e4dad74fcb5328b6b786b0bf06e00
Signed-off-by: Lan Zou <lanz@codeaurora.org>