meta-qti-display: fix eCRM compilation issue
The Makefile.am is not being copied to the Makefile in the source
folder because the override in the source folder occurs after
do_fetch(). As a result, the old Makefile is used, causing a
compilation failure in the eCRM build.

To prevent this issue, copy Makefile.am into Makefile during
do_configure:prepend().

Change-Id: I6271fcf302df486337be63d3590b8b31489764df
Signed-off-by: Anand Tarakh <quic_atarakh@quicinc.com>