meta-qti-display: Add the system group to weston for the purpose of allowing weston process to access node "/dev/dma_heap/qcom,system"
The Weston process needs to access the “/dev/dma_heap/qcom,system" node. After SELinux is enabled, there is a {dac_override} problem. The reason is that the Weston process DAC permission is "root/root", but the default permission of the node is 0640(owner: system, group: system). It cannot access the node normally, causing Weston to run abnormally.

Change-Id: I9aa40960cb1464b13550feb63ff9cf207615ef69